Proof of concept, tests passing, Full BC and Stratigility decoupled.
