+++ composer.json
@@ -19,15 +19,15 @@
         }
     ],
     "require": {
-        "php":                             "~7.0",
-        "psr/http-message":                "^1.0",
-        "lcobucci/jwt":                    "^3.1",
-        "zendframework/zend-stratigility": "^1.1",
-        "dflydev/fig-cookies":             "^1.0"
+        "php":                 "~7.0",
+        "psr/http-message":    "^1.0",
+        "lcobucci/jwt":        "^3.1",
+        "dflydev/fig-cookies": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit":              "^5.0",
-        "zendframework/zend-diactoros": "^1.1"
+        "phpunit/phpunit":                 "^5.0",
+        "zendframework/zend-stratigility": "^1.1",
+        "zendframework/zend-diactoros":    "^1.1"
     },
     "autoload": {
         "psr-4": {
